+++ extas.storage.php
@@ -1,0 +1,18 @@
+<?php
+
+return [
+  "name" => "jeyroik/extas-secrets",
+  "tables" => [
+    "secrets" => [
+      "namespace" => "extas\\repositories",
+      "item_class" => "extas\\components\\secrets\\Secret",
+      "pk" => "id",
+      "aliases" => [],
+      "hooks" => [],
+      "code" => [
+          'create-before' => '\\' . RepoItem::class . '::setId($item);'
+                            .'\\' . RepoItem::class . '::throwIfExist($this, $item, [\'name\']);'
+      ]
+    ]
+  ]
+];