+++ src/interfaces/secrets/ISecret.php
@@ -3,8 +3,9 @@ namespace extas\interfaces\secrets;
 
 use extas\interfaces\IDispatcherWrapper;
 use extas\interfaces\IHasValue;
+use extas\interfaces\IHaveUUID;
 use extas\interfaces\IItem;
-use extas\interfaces\samples\parameters\IHasSampleParameters;
+use extas\interfaces\parameters\IHaveParams;
 
 /**
  * Interface ISecret
@@ -12,13 +13,9 @@ use extas\interfaces\samples\parameters\IHasSampleParameters;
  * @package extas\interfaces\secrets
  * @author jeyroik <jeyroik@gmail.com>
  */
-interface ISecret extends IItem, IDispatcherWrapper, IHasSampleParameters, IHasValue
+interface ISecret extends IItem, IHaveUUID, IDispatcherWrapper, IHaveParams, IHasValue
 {
     public const SUBJECT = 'extas.secret';
-
-    public const FLAG__ENCRYPT = 'enc';
-    public const FLAG__DECRYPT = 'dec';
-    public const FIELD__TARGET = 'target';
 
     /**
      * @return bool
@@ -29,15 +26,4 @@ interface ISecret extends IItem, IDispatcherWrapper, IHasSampleParameters, IHasV
      * @return bool
      */
     public function decrypt(): bool;
-
-    /**
-     * @return string
-     */
-    public function getTarget(): string;
-
-    /**
-     * @param string $target
-     * @return $this
-     */
-    public function setTarget(string $target);
 }