+++ tests/secrets/SecretTest.php
@@ -1,11 +1,11 @@
 <?php
 namespace tests\secrets;
 
-use Dotenv\Dotenv;
+use extas\components\secrets\resolvers\Base64SecretResolver;
 use extas\components\secrets\Secret;
-use extas\interfaces\samples\parameters\ISampleParameter;
-use PHPUnit\Framework\TestCase;
-use tests\secrets\misc\ExampleResolver;
+use extas\interfaces\parameters\IParam;
+use tests\ExtasTestCase;
+use tests\resources\ExampleResolver;
 
 /**
  * Class SecretTest
@@ -13,26 +13,20 @@ use tests\secrets\misc\ExampleResolver;
  * @package tests\secrets
  * @author jeyroik <jeyroik@gmail.com>
  */
-class SecretTest extends TestCase
+class SecretTest extends ExtasTestCase
 {
-    protected function setUp(): void
-    {
-        parent::setUp();
-        $env = Dotenv::create(getcwd() . '/tests/');
-        $env->load();
-    }
+    protected array $libsToInstall = [
+        '' => ['php', 'php']
+        //'vendor/lib' => ['php', 'json'] storage ext, extas ext
+    ];
+    protected bool $isNeedInstallLibsItems = false;
+    protected string $testPath = __DIR__;
 
     public function testResolve()
     {
         $secret = new Secret([
-            Secret::FIELD__CLASS => ExampleResolver::class,
-            Secret::FIELD__VALUE => 'test',
-            Secret::FIELD__PARAMETERS => [
-                ExampleResolver::FIELD__RETURN => [
-                    ISampleParameter::FIELD__NAME => ExampleResolver::FIELD__RETURN,
-                    ISampleParameter::FIELD__VALUE => true
-                ]
-            ]
+            Secret::FIELD__CLASS => Base64SecretResolver::class,
+            Secret::FIELD__VALUE => 'test'
         ]);
 
         $encrypted = $secret->encrypt();
@@ -45,19 +39,11 @@ class SecretTest extends TestCase
         );
 
         $decrypted = $secret->decrypt();
-        $this->assertTrue($encrypted, 'Incorrect decrypting');
+        $this->assertTrue($decrypted, 'Incorrect decrypting');
         $this->assertEquals(
             'test',
             $secret->getValue(),
             'Incorrect value: ' . print_r($secret->getValue(), true)
         );
-
-        $secret->setParameterValue(ExampleResolver::FIELD__RETURN, false);
-
-        $resolved = $secret->encrypt();
-        $this->assertFalse($resolved, 'Incorrect exception catching');
-
-        $secret->setTarget('test');
-        $this->assertEquals('test', $secret->getTarget(), 'Incorrect target');
     }
 }