+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2.0.0
+
+- Updated package to `extas-foundation v6`.
+- Removed `ISecret::FIELD__TARGET`.
+- Added `ISecret::FIELD__ID`.
+- `ISecret` extends `extas\interfaces\parameters\IHaveParams` instead of `extas\interfaces\samples\parameters\IHasSampleParameters` now.
+- Added enum `ESecretFlag` instead of consts `ISecret::FLAG__ENCRYPT`, `ISecret::FLAG__DECRYPT`.
+- Added `Base64SecretResolver` as basic secret resolver.
+
 # 1.0.0
 
 - Secret methods `ecrypt` and `decrypt` added. You shoud use them instead of `resolve` in the previous version.